JAMES-3056 Merge MailboxAssert and MailboxSoftlyAssert into one
